Bunch of debugging and fixing to make FOC work again

+++ Inc/defines.h
@@ -147,7 +147,7 @@ typedef struct {
     uint16_t angle;		// Angle (or angle advance)
     uint16_t angle_min;
     uint16_t angle_max;
-    int8_t speed;		// Speed in angle increments per modulator call for FOC
+    int16_t speed;		// Speed in angle increments per modulator call for FOC
     uint8_t enable;
   } ctrl;
 } motor_state_t;
